New experiment
feat: Start new Jupyter notebook for two-nets idea

In this notebook, an attempt to create better Traffic Sign classifier
was made. Basic idea: train two CNN: one - for traffic sign
classification in general, and second - for numbers on speed limit sign
classification. It should work better then only one CNN, which was used
originally for the project as it was observed, that it usually wrong
classify speed limits, but is correct about sign type.

This notebook uses Keras. Inception module implemented in Keras as well.